+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## [Unreleased]
+### Fixed
+- Record Creations - Prevent issue on record creations for models with UUID primary key default value.
 
 ## RELEASE 1.1.17 - 2017-08-09
 ### Added